Update concatVerticalTest.cpp
Update concatVerticalTest.cpp

+++ test/concatVerticalTest.cpp
@@ -20,7 +20,7 @@ void concatVerticalTest()
 {
     TinyDIP::Image<TinyDIP::GrayScale> image1(3, 3);
     image1.setAllValue(10);
-    image1.at(2, 2) = 1;
+    image1.at(static_cast<std::size_t>(2), static_cast<std::size_t>(2)) = 1;
     std::cout << "Width: " + std::to_string(image1.getWidth()) + "\n";
     std::cout << "Height: " + std::to_string(image1.getHeight()) + "\n";
     auto image2 = TinyDIP::copyResizeBicubic(image1, 12, 12);
